requirements.txt
google-generativeai
streamlit
pillow 
python-dotenv
speechrecognition 
pydub
streamlit_webrtc
audio-recorder-streamlit